requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
keyboard==0.13.5
langchain==0.2.0
openai==1.30.1
selenium==4.21.0
weaviate_client==3.22.1
tiktoken==0.7.0
langchain-community==0.2.0
pandas==2.0.3
openpyxl==3.1.2
xlsxwriter==3.1.2
sentence-transformers==2.7.0
pydantic==2.7.1
transformers
datasets
